pyproject.toml
[tool.poetry]
name = "unblock-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = {version = "^0.23.2", extras = ["standard"]}
openai = "^0.28.0"
motor = "^3.3.1"
jsonschema = "^4.19.0"


[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
python-dotenv = "^1.0.0"
httpx = "^0.25.0"
pytest = "^7.4.2"
pytest-asyncio = "^0.21.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
